Gan xong Battle manager

+++ Assets/Scripts/Skill.cs
@@ -26,7 +26,7 @@ public class Skill : ScriptableObject
         }
         else { return this.skillDescription.upgrade; }
     }
-    public float GetSkillEffectAmount()
+    public int GetSkillEffectAmount()
     {
         if (!isUpgrade) { return this.effectAmount.nonUpgrade; }
         else { return this.effectAmount.upgrade;}
@@ -45,8 +45,8 @@ public struct SkillDescription
 //Luong dmg gay ra || Heal || Shield neu can
 public struct SkillEffectAmount
 {
-    public float nonUpgrade;
-    public float upgrade;
+    public int nonUpgrade;
+    public int upgrade;
 }
 [System.Serializable]
